Added filter.txt to start collecting banned phrases.
If we are unable to load the names from the names.txt, we default to using Stranger 1 and Stranger 2.

Implemented banned phrase filtering.

Added filter toggling to the main window.

Fixed an issue introduced in refactoring that caused toggling the block messages check box to enter an infinite loop.

Added the logic responsible for displaying the recaptcha window when challenged.

We no longer attempt to submit a recaptcha response unless the user has actually entered some text.

Finished removal of deprecated code.

Updated the icon such that it is more visible.

Set the size for the recaptcha window in the yml file. It is too big, and it needs to be tuned before release.

Included license text in project root.

Included header text in project root.

Removed deprecated dependencies from pom.

Disabled proguard execution. This is probably a precursor to its removal.